pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gcpx"
version = "0.1.0"
description = "A Python decorator to push data to Google Cloud Platform"
authors = [{ name = "mmngreco", email = "maxgreco@gmail.com" }]
readme = "README.md"
requires-python = ">=3.6"
dependencies = [
    "google-cloud-storage",
    "pandas",
    "google-auth",
    "pyarrow",
]

